+++ libs/aitools/agents/detect_test.go
@@ -218,85 +218,3 @@ func TestOpenCodeConfigDir(t *testing.T) {
 		assert.Equal(t, filepath.Join(home, ".config", "opencode"), dir)
 	})
 }
-
-func TestClaudePluginVersion(t *testing.T) {
-	tests := []struct {
-		name string
-		// manifest is the installed_plugins.json body, or "" to write no file.
-		manifest    string
-		wantVersion string
-		wantOK      bool
-	}{
-		{
-			name:   "no manifest file",
-			wantOK: false,
-		},
-		{
-			name:        "plugin installed",
-			manifest:    `{"version":2,"plugins":{"databricks@claude-plugins-official":[{"version":"0.2.9"}]}}`,
-			wantVersion: "0.2.9",
-			wantOK:      true,
-		},
-		{
-			// One entry per install scope; the highest version wins (not lexical).
-			name:        "highest version across scopes",
-			manifest:    `{"version":2,"plugins":{"databricks@claude-plugins-official":[{"version":"0.2.9"},{"version":"0.2.10"},{"version":"0.2.8"}]}}`,
-			wantVersion: "0.2.10",
-			wantOK:      true,
-		},
-		{
-			name:        "marketplace segment varies",
-			manifest:    `{"version":2,"plugins":{"databricks@isaac-sync-official":[{"version":"0.2.9"}]}}`,
-			wantVersion: "0.2.9",
-			wantOK:      true,
-		},
-		{
-			name:     "different plugin id does not match",
-			manifest: `{"version":2,"plugins":{"gopls-lsp@claude-plugins-official":[{"version":"1.0.0"}]}}`,
-			wantOK:   false,
-		},
-		{
-			name:     "malformed manifest",
-			manifest: `not json`,
-			wantOK:   false,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			cfg := t.TempDir()
-			if tc.manifest != "" {
-				require.NoError(t, os.MkdirAll(filepath.Join(cfg, "plugins"), 0o755))
-				require.NoError(t, os.WriteFile(filepath.Join(cfg, "plugins", "installed_plugins.json"), []byte(tc.manifest), 0o644))
-			}
-			a := &Agent{
-				Name:          "claude-code",
-				Plugin:        &PluginSpec{ID: "databricks"},
-				ConfigDir:     func(_ context.Context) (string, error) { return cfg, nil },
-				pluginVersion: claudePluginVersion,
-			}
-			version, ok := a.DatabricksPluginVersion(t.Context())
-			assert.Equal(t, tc.wantOK, ok)
-			assert.Equal(t, tc.wantVersion, version)
-		})
-	}
-}
-
-func TestDatabricksPluginVersionWithoutReader(t *testing.T) {
-	configDir := func(_ context.Context) (string, error) { return t.TempDir(), nil }
-
-	t.Run("no plugin spec", func(t *testing.T) {
-		a := &Agent{Name: "cursor", ConfigDir: configDir}
-		_, ok := a.DatabricksPluginVersion(t.Context())
-		assert.False(t, ok)
-	})
-
-	t.Run("no verified manifest reader", func(t *testing.T) {
-		// An agent with a plugin but no pluginVersion fn reports no version, so
-		// callers fall back to the CLI install state rather than guessing its
-		// on-disk layout.
-		a := &Agent{Name: "codex", Plugin: &PluginSpec{ID: "databricks"}, ConfigDir: configDir}
-		_, ok := a.DatabricksPluginVersion(t.Context())
-		assert.False(t, ok)
-	})
-}
